Resolve non-state-changed pages only
Only pages whose state has not change since a fetch was called, will be
resolved.

A possible edge case scenario:
A fetch is made to a page index, and before the request is resolved, this page in unloaded by a
setReadOffset. In that case, we do not resolve the page, and we discard
the records.

Add tests for exactly the edge case above.