Extract ember-impagination core to an  NPM package
This follows the template for es6 NPM packages laid out by
https://github.com/benjamn/jsnext-skeleton for distribution of npm
packages that contain both ES6 and commonjs modules.

The touch is super-lightweight, and it's much more clear what's going
on. Source files are in `src/`, where they can be included into an es6
build like `ember-cli`'s. To run the tests:

```
$ npm test
```

The tests are still failing, primarily because they depend on ember-cli
mirage, so we need to figure out a way to mock the server that doesn't
depend on it. Maybe using Faker directly?