Fix load / unload of pages
Load the page the moment the end of the load horizon "breaches" the
first record of the page when incrementing.
Load the page the moment the beginning of the load horizon "breaches"
the last record of the page when decrementing.
Unload the page the moment the beginning of the unload horizon "passes" the
last record of the page when incrementing.
Unload the page the moment the end of the unload horizon "passes"
the first record of the page when decrementing.