Add `util` to the dependencies for impagination
binary-search-tree requires `util`, but does not enumerate it as a
dependency. This is a known issue with it
https://github.com/louischatriot/node-binary-search-tree/issues/14

This adds `util` as a primary dependency so that this require will not
fail. While we're fixing up meta-information, this adds build time
artifacts  `/lib` and `/yarn.lock` to the ignore list.